# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "List of books documentation API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/addindb/": {
            "post": {
                "tags": [
                    "CrudController"
                ],
                "summary": "Create book",
                "operationId": "973e1c6446536102365300e9a78292ee",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Some author"
                                            },
                                            "published_year": {
                                                "type": "integer",
                                                "example": 2025
                                            },
                                            "isbn": {
                                                "type": "string",
                                                "example": "Some isbn"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "published_year": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "isbn": {
                                                    "type": "string",
                                                    "example": "Some isbn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/": {
            "get": {
                "tags": [
                    "CrudController"
                ],
                "summary": "Read book",
                "operationId": "766b6f8075c23d95bc8658adb428324c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "author": {
                                                "type": "string",
                                                "example": "Some author"
                                            },
                                            "published_year": {
                                                "type": "integer",
                                                "example": 2025
                                            },
                                            "isbn": {
                                                "type": "string",
                                                "example": "Some isbn"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Some title"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "Some author"
                                                    },
                                                    "published_year": {
                                                        "type": "integer",
                                                        "example": 2025
                                                    },
                                                    "isbn": {
                                                        "type": "string",
                                                        "example": "Some isbn"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books.delete/{id}/": {
            "get": {
                "tags": [
                    "CrudController"
                ],
                "summary": "Delete book",
                "operationId": "96b8a7c345a4c6cd03b735adce1d21a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID book",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "published_year": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "isbn": {
                                                    "type": "string",
                                                    "example": "Some isbn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books.formedit/{id}/": {
            "get": {
                "tags": [
                    "CrudController"
                ],
                "summary": "Show form update book",
                "operationId": "b5385d833048eaa2411b3202db3e0e27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID book",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "published_year": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "isbn": {
                                                    "type": "string",
                                                    "example": "Some isbn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/search/": {
            "post": {
                "tags": [
                    "SearchController"
                ],
                "summary": "Search book",
                "operationId": "0860cdae0f6895a2f168e5f3284f5f8d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "author": {
                                                "type": "string",
                                                "example": "Some author"
                                            },
                                            "published_year": {
                                                "type": "integer",
                                                "example": 2025
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author": {
                                                    "type": "string",
                                                    "example": "Some author"
                                                },
                                                "published_year": {
                                                    "type": "integer",
                                                    "example": 2025
                                                },
                                                "isbn": {
                                                    "type": "string",
                                                    "example": "Some isbn"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}